inserimento della paginazione ed ottimizzazione di jUnit

+++ src/main/java/com/example/demo/controllers/ProductController.java
@@ -1,5 +1,7 @@
 package com.example.demo.controllers;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,51 +17,58 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.demo.entities.Prodotti;
 import com.example.demo.services.ProdottiServices;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+
 
 @RestController
-@RequestMapping("/prodotti") 
+@RequestMapping("/prodotti")
+@Tag(name = "Prodotti Controller", description = "Gestione dei prodotti") // Titolo del controller
 public class ProductController {
 		
 		@Autowired
 		private ProdottiServices prodottiServices;
-		
-		@GetMapping(value = "/message")
-		public String getMessage() {
-			return "Hello from the server! Spring Boot";
-		}
-		
+				
 		@GetMapping(value = "/getAll", produces = "application/json")
+		@Operation(summary = "Recupera tutti i prodotti", description = "Restituisce un elenco di tutti i prodotti disponibili.")
 		public ResponseEntity<Iterable<Prodotti>> getProdotti() {
 			Iterable<Prodotti> prodotti = prodottiServices.findAll();
 			return new ResponseEntity<>(prodotti, HttpStatus.OK);
 		}		
 		
 		@GetMapping(value = "getProdottiById/{id}", produces = "application/json") 
-		public ResponseEntity<Prodotti> getProdottiById(@PathVariable("id") Integer id) {
-			Prodotti prodotti = prodottiServices.findById(id);
+		@Operation(summary = "Recupera un prodotto", description = "Restituisce un prodotto in base all'id.")
+		public ResponseEntity <Prodotti> getProdottiById(@PathVariable("id") Integer id) {
+			
+			Optional<Prodotti> prodotti = prodottiServices.findById(id);
 			 
 			    if (prodotti != null) {
-		            return new ResponseEntity<>(prodotti, HttpStatus.OK);
-		        } else {
-		            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		             return new ResponseEntity<>(prodotti.get(), HttpStatus.OK); 
+			    } else {
+		             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		        }
 		 }
 		
 		@PostMapping(value = "/create", consumes = "application/json", produces = "application/json")
-	    public ResponseEntity<Prodotti> createProdotto(@RequestBody Prodotti prodotto) {
+	    @Operation(summary = "Crea un nuovo prodotto", description = "Crea un nuovo prodotto.")
+		public ResponseEntity<Prodotti> createProdotto(@RequestBody Prodotti prodotto) {
 			Prodotti nuovoProdotto = prodottiServices.saveProdotti(prodotto);
 	        return new ResponseEntity<>(nuovoProdotto, HttpStatus.CREATED);
 	    }
 		
 		@DeleteMapping(value = "/delete/{id}")
-		public ResponseEntity<HttpStatus> deleteProdotto(@PathVariable("id") Integer id) {
+		@Operation(summary = "Elimina un prodotto", description = "Elimina un prodotto in base all'id.")
+		public ResponseEntity<HttpStatus> deleteProdotto(@PathVariable("id") int id) {
+			//Verifica se il prodotto esiste
+		System.out.println("id: " +id);			
+		System.out.println("prodottiServices.existsById(id)" + prodottiServices.existsById(id));	
+		
+		    if (!prodottiServices.existsById(id)) {
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			}
+			
 			prodottiServices.deleteProdotti(id);
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 		}
-		
-		
-		
-		
-		
 	
 }
